#390 Refactor large lib.rs files exceeding 2000 lines
